requirements.txt
accelerate==0.33.0
accelerate==0.31.0
bitsandbytes==0.43.3
braceexpand==0.1.7
diffusers==0.29.0
einops==0.8.0
h5py==3.11.0
huggingface_hub==0.23.4
imageio==2.34.2
matplotlib==3.7.5
numpy==1.24.4
openai_clip==1.0.1
opencv_python==4.10.0.84
openTSNE==1.0.2
packaging==24.1
Pillow==10.4.0
ptlflow==0.3.2
scenedetect==0.6.4
scikit_learn==1.3.2
scipy==1.14.1
skimage==0.0
torch==2.4.0+cu124
torch==2.3.0
torchvision==0.19.0+cu124
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.2
wandb==0.18.1
webdataset==0.2.100